ft: rename json to parsed json
