+++ src/core/communication-strategies/communication-strategy.ts
@@ -1,4 +1,4 @@
-import { ParsedJSON } from "../../common/json";
+import { ParsedJSON } from "../../common/parsed-json";
 import { PayloadForRequestHandler } from "./payloads/payload-for-request-handler";
 import { Response } from "../response";
