+++ src/index.ts
@@ -5,7 +5,7 @@ import { RequestMapper } from "./core/request-mapper";
 import { readFile } from "fs/promises";
 import path from "path";
 import { HTTPMethods } from "./core/http-methods";
-import { ParsedJSON } from "./common/json";
+import { ParsedJSON } from "./common/parsed-json";
 
 const server = fastify();
