+++ src/kafka/message.ts
@@ -1,4 +1,4 @@
-import { ParsedJSON } from "../common/json";
+import { ParsedJSON } from "../common/parsed-json";
 import { v4 as uuid } from "uuid";
 
 type StringObject = Record<string, string | number | boolean>;
